Latched entities check to see if the entity is spawned in the world else set dead. Catch throwable rather than exception when trying to recursively go through superclasses for setsize,hurt/death sounds. Should fix a crash with Morph.

+++ src/main/java/me/ichun/mods/ichunutil/client/core/EntityTrackerHandler.java
@@ -47,7 +47,7 @@ public class EntityTrackerHandler
         for(int i = latchedRendererEntities.size() - 1; i >= 0; i--) //Check the latched renderers
         {
             EntityLatchedRenderer latchedRenderer = latchedRendererEntities.get(i);
-            if(latchedRenderer.latchedEnt != null && (!latchedRenderer.latchedEnt.isDead || !latchedRenderer.latchedEnt.isEntityAlive() && latchedRenderer.maxDeathPersistTime > 0 && latchedRenderer.currentDeathPersistTime < latchedRenderer.maxDeathPersistTime)) //latched ent exists and is alive and well, or is persisting post-death
+            if(latchedRenderer.latchedEnt != null && (!latchedRenderer.latchedEnt.isDead && latchedRenderer.latchedEnt.addedToChunk || !latchedRenderer.latchedEnt.isEntityAlive() && latchedRenderer.maxDeathPersistTime > 0 && latchedRenderer.currentDeathPersistTime < latchedRenderer.maxDeathPersistTime)) //latched ent exists and is alive and well, or is persisting post-death
             {
                 if(latchedRenderer.isDead || (iChunUtil.eventHandlerClient.ticks - latchedRenderer.lastUpdate) > 10 && !Minecraft.getMinecraft().isGamePaused())//latcher died/stopped updating, kill it replace with new one.
                 {
